requirements.txt
# Core FastAPI web framework
uvicorn[standard]>=0.29.0
fastapi>=0.116.1
starlette>=0.47.2
jinja2>=3.1.3
python-multipart>=0.0.9

# Configuration and environment
python-dotenv>=1.0.1
pydantic>=2.11.7

# Security and authentication
bcrypt>=4.1.2,<4.2.0
passlib[bcrypt]>=1.7.4
itsdangerous

# HTTP client and networking
httpx>=0.27.0
requests>=2.28.0

# System monitoring and metrics
psutil>=5.9.0
prometheus-client>=0.17.0

# Task scheduling
apscheduler>=3.10.4

# Notifications system
aiosmtplib>=3.0.0
email-validator>=2.0.0

# Discord Bot Integration
discord.py>=2.3.2

# Process identification
setproctitle>=1.3.0

# Core Python dependencies
typing-inspection>=0.4.1
typing_extensions>=4.14.1
anyio>=4.10.0
sniffio>=1.3.1
h11>=0.16.0
certifi>=2025.8.3
charset-normalizer>=3.4.2
idna>=3.10
MarkupSafe>=3.0.2
tzdata>=2025.2
tzlocal>=5.3.1
colorama>=0.4.6
